+++ src/vs/editor/common/services/textResourceConfigurationService.ts
