+++ src/vs/workbench/contrib/outline/browser/outlineNavigation.ts
@@ -19,7 +19,7 @@ import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegis
 import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
 import { OutlineFilter } from 'vs/editor/contrib/documentSymbols/outlineTree';
 import { binarySearch } from 'vs/base/common/arrays';
-import { ITextResourceConfigurationService } from 'vs/editor/common/services/resourceConfiguration';
+import { ITextResourceConfigurationService } from 'vs/editor/common/services/textResourceConfigurationService';
 
 class FlatOutline {
